Tweaking Folder Model display

+++ src/SonOfPicasso.Core/Services/ImageContainerManagementService.cs
@@ -1,10 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Reactive;
+using System.Reactive.Disposables;
 using System.Reactive.Linq;
+using System.Reactive.Subjects;
 using DynamicData;
 using Serilog;
+using SonOfPicasso.Core.Extensions;
 using SonOfPicasso.Core.Interfaces;
 using SonOfPicasso.Core.Model;
 using SonOfPicasso.Core.Scheduling;
@@ -12,19 +16,19 @@ using SonOfPicasso.Data.Model;
 
 namespace SonOfPicasso.Core.Services
 {
-    public class ImageContainerImageManagementService : IImageContainerManagementService
+    public class ImageContainerManagementService : IImageContainerManagementService
     {
+        private readonly Subject<IObservable<FileSystemEventArgs>> _currentFileWatcherSubject;
+        private readonly CompositeDisposable _disposables;
         private readonly IFolderRulesManagementService _folderRulesManagementService;
         private readonly IFolderWatcherService _folderWatcherService;
-
         private readonly SourceCache<IImageContainer, string> _imageContainerCache;
         private readonly IImageContainerOperationService _imageContainerOperationService;
         private readonly IObservableCache<ImageRef, string> _imageRefCache;
         private readonly ILogger _logger;
         private readonly ISchedulerProvider _schedulerProvider;
-        private IDisposable _folderManagementDisposable;
 
-        public ImageContainerImageManagementService(
+        public ImageContainerManagementService(
             IImageContainerOperationService imageContainerOperationService,
             IFolderWatcherService folderWatcherService,
             IFolderRulesManagementService folderRulesManagementService,
@@ -36,12 +40,19 @@ namespace SonOfPicasso.Core.Services
             _folderRulesManagementService = folderRulesManagementService;
             _schedulerProvider = schedulerProvider;
             _logger = logger;
-            _imageContainerCache = new SourceCache<IImageContainer, string>(imageContainer => imageContainer.Id);
+            _imageContainerCache = new SourceCache<IImageContainer, string>(imageContainer => imageContainer.Key);
             _imageRefCache = _imageContainerCache
                 .Connect()
                 .ObserveOn(schedulerProvider.TaskPool)
-                .TransformMany(container => container.ImageRefs, imageRef => imageRef.Id)
+                .TransformMany(container => container.ImageRefs, imageRef => imageRef.Key)
                 .AsObservableCache();
+
+            _disposables = new CompositeDisposable();
+
+            _currentFileWatcherSubject = new Subject<IObservable<FileSystemEventArgs>>();
+            _currentFileWatcherSubject.Switch()
+                .Subscribe(HandlerFolderWatcherEvent)
+                .DisposeWith(_disposables);
         }
 
         public IConnectableCache<IImageContainer, string> ImageContainerCache => _imageContainerCache;
@@ -50,9 +61,35 @@ namespace SonOfPicasso.Core.Services
 
         public void Dispose()
         {
-            _folderManagementDisposable?.Dispose();
             _imageContainerCache?.Dispose();
             _imageRefCache?.Dispose();
+            _currentFileWatcherSubject?.Dispose();
+            _disposables?.Dispose();
+        }
+
+        public IObservable<Unit> ResetRules(IEnumerable<FolderRule> folderRules)
+        {
+            var folderRulesArray = folderRules.ToArray();
+
+            var resetRules = _folderRulesManagementService.ResetFolderManagementRules(folderRulesArray);
+
+            var applyImageChanges = _imageContainerOperationService.ApplyRuleChanges(folderRulesArray)
+                .Select(changes =>
+                {
+                    _imageContainerCache.Remove(changes.DeletedFolderIds.Select(FolderImageContainer.GetContainerId));
+
+                    return Unit.Default;
+                });
+
+            var result = resetRules
+                .Zip(applyImageChanges, (unit, _) => unit);
+
+            return result;
+        }
+
+        public IObservable<ResetChanges> PreviewResetRulesChanges(IEnumerable<FolderRule> folderRules)
+        {
+            return _imageContainerOperationService.PreviewRuleChangesEffect(folderRules);
         }
 
         public IObservable<Unit> Start()
@@ -66,8 +103,7 @@ namespace SonOfPicasso.Core.Services
 
         public void Stop()
         {
-            _folderManagementDisposable?.Dispose();
-            _folderManagementDisposable = null;
+            _currentFileWatcherSubject.OnNext(Observable.Never<FileSystemEventArgs>());
         }
 
         public IObservable<IImageContainer> ScanFolder(string path)
@@ -123,44 +159,45 @@ namespace SonOfPicasso.Core.Services
         public IObservable<Unit> DeleteAlbum(int albumId)
         {
             return _imageContainerOperationService.DeleteAlbum(albumId)
-                .Do(container => _imageContainerCache.Remove(AlbumImageContainer.GetContainerId(albumId)));
+                .Do(container => _imageContainerCache.Remove(AlbumImageContainer.GetContainerKey(albumId)));
         }
 
         private void StartWatcher()
         {
-            _folderManagementDisposable = _folderRulesManagementService.GetFolderManagementRules()
-                .SelectMany(list => _folderWatcherService.WatchFolders(list))
-                .Subscribe(HandlerFolderWatcherEvent);
+            var observable = _folderRulesManagementService.GetFolderManagementRules()
+                .SelectMany(list => _folderWatcherService.WatchFolders(list, Constants.ImageExtensions));
+
+            _currentFileWatcherSubject.OnNext(observable);
         }
 
         private void HandlerFolderWatcherEvent(FileSystemEventArgs args)
         {
-            _logger.Debug("HandlerFolderWatcherEvent {ChangeType} {Name}", args.ChangeType, args.Name);
-
-            switch (args.ChangeType)
+            if (args.ChangeType != WatcherChangeTypes.Created)
             {
-                case WatcherChangeTypes.Created:
-                    break;
+                _logger.Debug("HandlerFolderWatcherEvent {ChangeType} {Name}", args.ChangeType, args.Name);
 
-                case WatcherChangeTypes.Changed:
-                    _imageContainerCache.PopulateFrom(
-                        _imageContainerOperationService.AddOrUpdateImage(args.FullPath));
-                    break;
+                switch (args.ChangeType)
+                {
+                    case WatcherChangeTypes.Changed:
+                        _imageContainerCache.PopulateFrom(
+                            _imageContainerOperationService.AddOrUpdateImage(args.FullPath));
+                        break;
 
-                case WatcherChangeTypes.Deleted:
-                    _imageContainerCache.PopulateFrom(
-                        _imageContainerOperationService.DeleteImage(args.FullPath));
-                    break;
+                    case WatcherChangeTypes.Deleted:
+                        _imageContainerCache.PopulateFrom(
+                            _imageContainerOperationService.DeleteImage(args.FullPath));
+                        break;
 
-                case WatcherChangeTypes.Renamed:
-                    var renamedEventArgs = (RenamedEventArgs) args;
-                    _imageContainerCache.PopulateFrom(
-                        _imageContainerOperationService.RenameImage(renamedEventArgs.OldFullPath,
-                            renamedEventArgs.FullPath));
-                    break;
+                    case WatcherChangeTypes.Renamed:
+                        var renamedEventArgs = (RenamedEventArgs) args;
+                        _imageContainerCache.PopulateFrom(
+                            _imageContainerOperationService.RenameImage(renamedEventArgs.OldFullPath,
+                                renamedEventArgs.FullPath));
+                        break;
 
-                default:
-                    throw new ArgumentOutOfRangeException();
+                    default:
+                        throw new ArgumentOutOfRangeException();
+                }
             }
         }
     }
